30.11.21

+++ src/main/java/com/kon/EShop/util/FileManager.java
@@ -2,6 +2,8 @@ package com.kon.EShop.util;
 
 import com.kon.EShop.DelLabel;
 import com.kon.EShop.model.MyFile;
+import lombok.extern.java.Log;
+import org.imgscalr.Scalr;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 import org.springframework.util.ResourceUtils;
@@ -10,6 +12,7 @@ import org.springframework.web.multipart.MultipartFile;
 import javax.imageio.ImageIO;
 import javax.swing.*;
 import java.awt.*;
+import java.awt.geom.AffineTransform;
 import java.awt.geom.Rectangle2D;
 import java.awt.image.BufferedImage;
 import java.io.*;
@@ -17,11 +20,11 @@ import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
+import java.util.*;
 import java.util.List;
 import java.util.stream.Collectors;
 
+@Log
 @Component
 public class FileManager {
 
@@ -34,23 +37,38 @@ public class FileManager {
     @Value("${upload.photo}")
     private String photo;
 
+    @Value("${main.category.height}")
+    private int mcHeight;
+
+    @Value("${main.category.width}")
+    private int mcWidth;
+
+    @Value("${catalog.photo.height}")
+    private int catalogHeight;
+
+    @Value("${catalog.photo.width}")
+    private int catalogWidth;
+
+    @Value("${upload.small}")
+    private String small;
+
     public String upload(MultipartFile resource, String keyName, String where) throws IOException {
         Files.createDirectories(Paths.get(where));
-
         Path file = Files.createFile(Paths.get(where, keyName));
-        resource.transferTo(file);
+        String name = keyName.substring(keyName.lastIndexOf("-") + 1);
+        saveLabelOfPhoto(where, resource, mcWidth, mcHeight, name);
+
         return file.toString();
     }
 
     public void delete(String key, String where) throws IOException {
         if (!key.equals("temp.jpg") && !key.equals("favicon.ico")) {
-            Path path = Paths.get(this.path + (where.equals("label") ?
-                    label :
-                    where.equals("csv") ?
-                            getAbsolutePath() + "static/csv/" :
-                            photo) + key);
-
-            Files.delete(path);
+            Path path = Paths.get(this.path + (!where.equals("label") ? where.equals("csv") ? getAbsolutePath() + "static/csv/" : photo : label) + key);
+            if (Files.exists(path)) Files.delete(path);
+            if (where.equals("photos")) {
+                path = Paths.get(this.path + small + key);
+                if (Files.exists(path)) Files.delete(path);
+            }
         }
     }
 
@@ -62,7 +80,7 @@ public class FileManager {
                 if (dir.equals("label"))
                     upload(fileData, name, path + label);
                 if (dir.equals("photos"))
-                    waterMark("AutoMir", path + photo, fileData);
+                    waterMark("am.dn.ua", path + photo, fileData);
                 uploaded.add(name);
             }
         }
@@ -122,7 +140,7 @@ public class FileManager {
     public void waterMark(String mark, String path, MultipartFile file) throws IOException {
         ImageIcon image = new ImageIcon(file.getBytes());
         BufferedImage bufferedImage = new BufferedImage(image.getIconWidth(), image.getIconHeight(), BufferedImage.TYPE_INT_BGR);
-        Graphics2D g2d = (Graphics2D) bufferedImage.getGraphics();
+        Graphics2D g2d = bufferedImage.createGraphics();
         g2d.drawImage(image.getImage(), 0, 0, null);
         AlphaComposite alpha = AlphaComposite.getInstance(AlphaComposite.SRC_OVER, 0.5f);
         g2d.setComposite(alpha);
@@ -138,9 +156,44 @@ public class FileManager {
         g2d.dispose();
 
         String name = file.getOriginalFilename();
-        OutputStream out = new PrintStream(path + name);
-        ImageIO.write(bufferedImage, "jpg", out);
-        out.close();
+        saveImage(bufferedImage, path + name);
+        saveLabelOfPhoto(this.path + this.small, file, catalogWidth, catalogHeight, null);
+    }
+
+    public void saveLabelOfPhoto(String path, MultipartFile file, int imgWidth, int imgHeight, String name) throws IOException {
+        ImageIcon image = new ImageIcon(file.getBytes());
+
+        int width = image.getIconWidth();
+        int height = image.getIconHeight();
+
+        int scaleW, scaleH;
+        int d = (int) (imgWidth * 1.0 / width * height);
+        if (d <= imgHeight) {
+            scaleW = imgWidth;
+            scaleH = d;
+        }else {
+            scaleH = imgHeight;
+            d = (int) (imgHeight * 1.0 / height * width +0.5);
+            scaleW = d;
+        }
+
+        BufferedImage bufferedImage = resizeImage(ImageIO.read(file.getInputStream()), scaleW, scaleH);
+
+        if (name == null) name = file.getOriginalFilename();
+        saveImage(bufferedImage, path + name);
+    }
+
+    private BufferedImage resizeImage(BufferedImage originalImage, int targetWidth, int targetHeight) {
+        long time = new Date().getTime();
+        BufferedImage resize = Scalr.resize(originalImage, targetWidth, targetHeight, Scalr.OP_ANTIALIAS);
+        log.info("Resizing image took " + (new Date().getTime() - time) + "ms");
+        return resize;
+    }
+
+    private void saveImage(BufferedImage bufferedImage, String fullName) throws IOException {
+        File out = new File(fullName);
+        String fileExtension = fullName.substring(fullName.lastIndexOf(".") + 1);
+        ImageIO.write(bufferedImage, fileExtension, out);
     }
 
 }
